Add type-hinting to return types
I noted that the CI suite only builds against PHP 7.1 and 7.2.  Therefore I assumed this package was aimed at recent versions of PHP only where type-hinting return types are supported.  I went and added the type-hinting as a way to help promote this more recent construct in PHP to help make the code more readable, and maintainable in the future.